authorization implemented

+++ src/api/notes/handler.js
@@ -16,8 +16,11 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentiaId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentiaId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -49,8 +52,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentiaId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentiaId);
     return {
       status: 'success',
       data: {
@@ -62,7 +66,11 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentiaId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentiaId);
       const note = await this._service.getNoteById(id);
+
       return {
         status: 'success',
         data: {
@@ -93,10 +101,11 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title, body, tags } = request.payload;
       const { id } = request.params;
+      const { id: credentiaId } = request.auth.credentials;
 
-      await this._service.editNoteById(id, { title, body, tags });
+      await this._service.verifyNoteOwner(id, credentiaId);
+      await this._service.editNoteById(id, request.payload);
 
       return {
         status: 'success',
@@ -126,6 +135,9 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentiaId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentiaId);
       await this._service.deleteNoteById(id);
 
       return {
